+++ pyccel/codegen/wrapper/c_to_python_wrapper.py
@@ -10,12 +10,13 @@ which creates an interface exposing C code to Python.
 import warnings
 from pyccel.ast.bind_c        import BindCFunctionDef, BindCPointer, BindCFunctionDefArgument
 from pyccel.ast.bind_c        import BindCModule, BindCVariable, BindCFunctionDefResult
+from pyccel.ast.bind_c        import BindCClassDef
 from pyccel.ast.builtins      import PythonTuple
 from pyccel.ast.core          import Interface, If, IfSection, Return, FunctionCall
 from pyccel.ast.core          import FunctionDef, FunctionDefArgument, FunctionDefResult
 from pyccel.ast.core          import Assign, AliasAssign, Deallocate, Allocate
 from pyccel.ast.core          import Import, Module, AugAssign, CommentBlock
-from pyccel.ast.core          import FunctionAddress, Declare
+from pyccel.ast.core          import FunctionAddress, Declare, ClassDef
 from pyccel.ast.cwrapper      import PyModule, PyccelPyObject, PyArgKeywords
 from pyccel.ast.cwrapper      import PyArg_ParseTupleNode, Py_None, PyClassDef
 from pyccel.ast.cwrapper      import py_to_c_registry, check_type_registry, PyBuildValueNode
@@ -23,6 +24,7 @@ from pyccel.ast.cwrapper      import PyErr_SetString, PyTypeError, PyNotImplemen
 from pyccel.ast.cwrapper      import C_to_Python, PyFunctionDef, PyInterface
 from pyccel.ast.cwrapper      import PyModule_AddObject, Py_DECREF, PyObject_TypeCheck
 from pyccel.ast.cwrapper      import Py_INCREF, PyType_Ready, WrapperCustomDataType
+from pyccel.ast.cwrapper      import PyccelPyTypeObject
 from pyccel.ast.c_concepts    import ObjectAddress, PointerCast
 from pyccel.ast.datatypes     import NativeVoid, NativeInteger, CustomDataType, DataTypeFactory
 from pyccel.ast.datatypes     import NativeNumeric
@@ -57,6 +59,8 @@ class CToPythonWrapper(Wrapper):
         self._python_object_map = {}
         # Indicate if arrays were wrapped.
         self._wrapping_arrays = False
+        # The object that should be returned to indicate an error
+        self._error_exit_code = Nil()
         super().__init__()
 
     def get_new_PyObject(self, name, dtype = None, is_temp = False):
@@ -121,11 +125,15 @@ class CToPythonWrapper(Wrapper):
         list of Variable
             Variables which will hold the arguments in Python.
         """
-        collect_args = [self.get_new_PyObject(getattr(a, 'original_function_argument_variable', a.var).name+'_obj') for a in args]
+        orig_args = [getattr(a, 'original_function_argument_variable', a.var) for a in args]
+        is_bound = [getattr(a, 'wrapping_bound_argument', a.bound_argument) for a in args]
+        collect_args = [self.get_new_PyObject(o_a.name+'_obj',
+                                              dtype = o_a.dtype if b else None)
+                        for a, b, o_a in zip(args, is_bound, orig_args)]
         self._python_object_map.update(dict(zip(args, collect_args)))
         return collect_args
 
-    def _unpack_python_args(self, args):
+    def _unpack_python_args(self, args, class_base = None):
         """
         Unpack the arguments received from Python into the expected Python variables.
 
@@ -138,6 +146,10 @@ class CToPythonWrapper(Wrapper):
         ----------
         args : iterable of FunctionDefArguments
             The expected arguments of the function.
+
+        class_base : DataType, optional
+            The DataType of the class which the method belongs to. In the case of a method
+            defined in a module this value is None.
 
         Returns
         -------
@@ -166,23 +178,29 @@ class CToPythonWrapper(Wrapper):
             return NULL;
         }
         """
+        has_bound_arg = class_base is not None
+        bound_arg = args[0] if has_bound_arg else None
+        args = args[int(has_bound_arg):]
         # Create necessary variables
-        func_args = [self.get_new_PyObject(n) for n in ("self", "args", "kwargs")]
+        func_args = [self.get_new_PyObject("self", class_base)] + [self.get_new_PyObject(n) for n in ("args", "kwargs")]
         arg_vars  = self._get_python_argument_variables(args)
         keyword_list_name = self.scope.get_new_name('kwlist')
+
+        if has_bound_arg:
+            self._python_object_map[bound_arg] = func_args[0]
 
         # Create the list of argument names
         arg_names = [getattr(a, 'original_function_argument_variable', a.var).name for a in args]
         keyword_list = PyArgKeywords(keyword_list_name, arg_names)
 
-        # Initialise optionals
-        body = [AliasAssign(py_arg, Py_None) for func_def_arg, py_arg in zip(args, arg_vars) if func_def_arg.has_default]
-
         # Parse arguments
         parse_node = PyArg_ParseTupleNode(*func_args[1:], args, arg_vars, keyword_list)
 
+        # Initialise optionals
+        body = [AliasAssign(py_arg, Py_None) for func_def_arg, py_arg in zip(args, arg_vars) if func_def_arg.has_default]
+
         body.append(keyword_list)
-        body.append(If(IfSection(PyccelNot(parse_node), [Return([Nil()])])))
+        body.append(If(IfSection(PyccelNot(parse_node), [Return([self._error_exit_code])])))
 
         return func_args, body
 
@@ -254,11 +272,6 @@ class CToPythonWrapper(Wrapper):
                                arguments = [FunctionDefArgument(Variable(dtype=PyccelPyObject(), name = 'o', memory_handling='alias'))],
                                results   = [FunctionDefResult(Variable(dtype=dtype, name = 'v', precision = prec))])
 
-            if raise_error:
-                message = LiteralString(f"Expected an argument of type {dtype} for argument {arg.name}")
-                python_error = FunctionCall(PyErr_SetString, [PyTypeError, message])
-                error_code = (python_error,)
-
             func_call = FunctionCall(func, [py_obj])
         else:
             dtype = str(dtype)
@@ -281,6 +294,11 @@ class CToPythonWrapper(Wrapper):
             # No error code required as the error is raised inside pyarray_check
 
             func_call = FunctionCall(check_func, [py_obj, type_ref, LiteralInteger(rank), flag])
+
+        if raise_error:
+            message = LiteralString(f"Expected an argument of type {dtype} for argument {arg.name}")
+            python_error = FunctionCall(PyErr_SetString, [PyTypeError, message])
+            error_code = (python_error,)
 
         return func_call, error_code
 
@@ -422,7 +440,7 @@ class CToPythonWrapper(Wrapper):
         function = PyFunctionDef(name = name, arguments = func_args, results = func_results,
                 body = [FunctionCall(PyErr_SetString, [PyNotImplementedError,
                                         LiteralString(error_msg)]),
-                        Return([Nil()])],
+                        Return([self._error_exit_code])],
                 scope = scope, original_function = original_function)
 
         self.scope.functions[name] = function
@@ -500,6 +518,221 @@ class CToPythonWrapper(Wrapper):
 
         return FunctionDef(func_name, [FunctionDefArgument(module_var)], [FunctionDefResult(result_var)], body,
                 scope = func_scope, is_static=True)
+
+    def _get_class_allocator(self, class_dtype, func = None):
+        """
+        Create the allocator for the class.
+
+        Create a function which will allocate the memory for the class instance. This
+        is equivalent to the `__new__` function.
+
+        Parameters
+        ----------
+        class_dtype : DataType
+            The datatype of the class being translated.
+
+        func : FunctionDef, optional
+            The function which provides a new instance of the class.
+
+        Returns
+        -------
+        PyFunctionDef
+            A function that can be called to create the class instance.
+        """
+        if func:
+            func_name = self.scope.get_new_name(f'{func.name}___wrapper')
+        else:
+            func_name = self.scope.get_new_name(f'{class_dtype.name}__new___wrapper')
+        func_scope = self.scope.new_child_scope(func_name)
+        self.scope = func_scope
+
+        self_var = Variable(dtype=PyccelPyTypeObject(), name=self.scope.get_new_name('self'),
+                              memory_handling='alias')
+        self.scope.insert_variable(self_var)
+        func_args = [self_var] + [self.get_new_PyObject(n) for n in ("args", "kwargs")]
+        func_args = [FunctionDefArgument(a) for a in func_args]
+
+        func_results = [FunctionDefResult(self.get_new_PyObject("result", is_temp=True))]
+
+        # Get the results of the PyFunctionDef
+        python_result_var = self.get_new_PyObject('result_obj', class_dtype)
+        scope = python_result_var.cls_base.scope
+        attribute = scope.find('instance', 'variables', raise_if_missing = True)
+        c_res = attribute.clone(attribute.name, new_class = DottedVariable, lhs = python_result_var)
+
+        body = [Allocate(python_result_var, shape=(), order=None, status='unallocated', like = self_var)]
+
+        if func:
+            body.append(AliasAssign(c_res, FunctionCall(func, ())))
+        else:
+            result_name = self.scope.get_new_name('result')
+            result = Variable(class_dtype, result_name)
+            body.append(Allocate(c_res, shape=(), order=None, status='unallocated',
+                         like = result))
+
+        body.append(Return([ObjectAddress(PointerCast(python_result_var, func_results[0].var))]))
+
+        self.exit_scope()
+
+        return PyFunctionDef(func_name, func_args, func_results,
+                             body, scope=func_scope, original_function = None)
+
+    def _get_class_initialiser(self, init_function, cls_dtype):
+        """
+        Create the constructor for the class.
+
+        Create a function which will initialise the class. This function creates
+        the `__new__` function to allocate the memory which stores the class
+        instance and calls the `__init__` function.
+
+        Parameters
+        ----------
+        init_function : FunctionDef
+            The `__init__` function in the translated class.
+
+        cls_dtype : DataType
+            The datatype of the class being translated.
+
+        Returns
+        -------
+        new_function : PyFunctionDef
+            A function that can be called to create the class instance.
+
+        init_function : PyFunctionDef
+            A function that can be called to create the class instance.
+        """
+        original_func = getattr(init_function, 'original_function', init_function)
+        func_name = self.scope.get_new_name(original_func.name+'_wrapper')
+        func_scope = self.scope.new_child_scope(func_name)
+        self.scope = func_scope
+        self._error_exit_code = PyccelUnarySub(LiteralInteger(1, precision=-2))
+        class_dtype = init_function.arguments[0].var.dtype
+
+        is_bind_c_function_def = isinstance(init_function, BindCFunctionDef)
+
+        # Handle un-wrappable functions
+        if any(isinstance(getattr(a, 'original_function_argument_variable', a.var), FunctionAddress) for a in init_function.arguments):
+            self.exit_scope()
+            warnings.warn("Functions with functions as arguments will not be callable from Python")
+            return self._get_untranslatable_function(func_name,
+                         func_scope, init_function,
+                         "Cannot pass a function as an argument")
+
+        # Add the variables to the expected symbols in the scope
+        for a in init_function.arguments:
+            func_scope.insert_symbol(a.var.name)
+        for a in getattr(init_function, 'bind_c_arguments', ()):
+            func_scope.insert_symbol(a.original_function_argument_variable.name)
+
+        # Get variables describing the arguments and results that are seen from Python
+        python_args = init_function.bind_c_arguments if is_bind_c_function_def else init_function.arguments
+
+        # Get variables describing the arguments and results that must be passed to the function
+        original_c_args = init_function.arguments
+
+        # Get the arguments of the PyFunctionDef
+        func_args, body = self._unpack_python_args(python_args, class_dtype)
+        func_args = [FunctionDefArgument(a) for a in func_args]
+
+        # Get the results of the PyFunctionDef
+        python_result_variable = Variable(NativeInteger(), self.scope.get_new_name(),
+                                          precision = -2, is_temp = True)
+
+        # Get the code required to extract the C-compatible arguments from the Python arguments
+        body += [l for a in python_args for l in self._wrap(a)]
+
+        # Get the arguments and results which should be used to call the c-compatible function
+        func_call_args = [self.scope.find(n.var.name, category='variables', raise_if_missing = True) for n in original_c_args]
+
+        # Call the C-compatible function
+        body.append(FunctionCall(init_function, func_call_args))
+
+        # Deallocate the C equivalent of any array arguments
+        # The C equivalent is the same variable that is passed to the function unless the target language is Fortran.
+        # In this case the function arguments are the data pointer and the shapes and strides, but the C equivalent
+        # is an ndarray.
+        for a in original_c_args:
+            orig_var = getattr(a, 'original_function_argument_variable', a.var)
+            if orig_var.is_ndarray:
+                v = self.scope.find(orig_var.name, category='variables', raise_if_missing = True)
+                if v.is_optional:
+                    body.append(If( IfSection(PyccelIsNot(v, Nil()), [Deallocate(v)]) ))
+                else:
+                    body.append(Deallocate(v))
+
+        # Pack the Python compatible results of the function into one argument.
+        func_results = [FunctionDefResult(python_result_variable)]
+        body.append(Return([LiteralInteger(0, precision=-2)]))
+
+        self.exit_scope()
+        for a in python_args:
+            if not a.bound_argument:
+                self._python_object_map.pop(a)
+
+        function = PyFunctionDef(func_name, func_args, func_results, body, scope=func_scope,
+                docstring = init_function.docstring, original_function = original_func)
+
+        self.scope.functions[func_name] = function
+        self._python_object_map[init_function] = function
+        self._error_exit_code = Nil()
+
+        return function
+
+    def _get_class_destructor(self, del_function, cls_dtype, wrapper_scope):
+        """
+        Create the destructor for the class.
+
+        Create a function which will act as a destructor for the class. This
+        function calls the `__del__` function and frees the memory allocated
+        to store the class instance.
+
+        Parameters
+        ----------
+        del_function : FunctionDef
+            The `__del__` function in the translated class.
+
+        cls_dtype : DataType
+            The datatype of the class being translated.
+
+        wrapper_scope : Scope
+            The scope for the wrapped version of the class.
+
+        Returns
+        -------
+        PyFunctionDef
+            A function that can be called to destroy the class instance.
+        """
+        original_func = getattr(del_function, 'original_function', del_function)
+        func_name = self.scope.get_new_name(original_func.name+'_wrapper')
+        func_scope = self.scope.new_child_scope(func_name)
+        self.scope = func_scope
+
+        # Add the variables to the expected symbols in the scope
+        for a in del_function.arguments:
+            func_scope.insert_symbol(a.var.name)
+        func_arg = self.get_new_PyObject('self', cls_dtype)
+
+        attribute = wrapper_scope.find('instance', 'variables')
+        c_obj = attribute.clone(attribute.name, new_class = DottedVariable, lhs = func_arg)
+
+        if isinstance(del_function, BindCFunctionDef):
+            body = [FunctionCall(del_function, [c_obj]),
+                    Deallocate(func_arg)]
+        else:
+            body = [FunctionCall(del_function, [c_obj]),
+                    Deallocate(c_obj),
+                    Deallocate(func_arg)]
+
+        self.exit_scope()
+
+        function = PyFunctionDef(func_name, [FunctionDefArgument(func_arg)], [], body, scope=func_scope,
+                original_function = original_func)
+
+        self.scope.functions[func_name] = function
+        self._python_object_map[del_function] = function
+
+        return function
+
     #--------------------------------------------------------------------------------------------------------------------------------------------
 
     def _wrap_Module(self, expr):
@@ -586,6 +819,15 @@ class CToPythonWrapper(Wrapper):
         # Add external functions for normal functions
         for f in expr.funcs:
             external_funcs.append(FunctionDef(f.name.lower(), f.arguments, f.results, [], is_header = True, scope = Scope()))
+
+        for c in expr.classes:
+            m = c.new_func
+            external_funcs.append(FunctionDef(m.name, m.arguments, m.results, [], is_header = True, scope = Scope()))
+            for m in c.methods:
+                external_funcs.append(FunctionDef(m.name, m.arguments, m.results, [], is_header = True, scope = Scope()))
+            for i in c.interfaces:
+                for f in i.functions:
+                    external_funcs.append(FunctionDef(f.name, f.arguments, f.results, [], is_header = True, scope = Scope()))
         pymod.external_funcs = external_funcs
 
         return pymod
@@ -622,6 +864,11 @@ class CToPythonWrapper(Wrapper):
         self.scope = func_scope
         original_funcs = expr.functions
         example_func = original_funcs[0]
+        possible_class_base = expr.get_user_nodes((ClassDef,))
+        if possible_class_base:
+            class_dtype = possible_class_base[0].class_type
+        else:
+            class_dtype = None
 
         # Add the variables to the expected symbols in the scope
         for a in getattr(example_func, 'bind_c_arguments', example_func.arguments):
@@ -629,7 +876,7 @@ class CToPythonWrapper(Wrapper):
 
         # Create necessary arguments
         python_args = getattr(example_func, 'bind_c_arguments', example_func.arguments)
-        func_args, body = self._unpack_python_args(python_args)
+        func_args, body = self._unpack_python_args(python_args, class_dtype)
 
         # Get python arguments which will be passed to FunctionDefs
         python_arg_objs = [self._python_object_map[a] for a in python_args]
@@ -657,7 +904,7 @@ class CToPythonWrapper(Wrapper):
             functions.append(wrapped_func)
         if_sections.append(IfSection(LiteralTrue(),
                     [FunctionCall(PyErr_SetString, [PyTypeError, "Unexpected type combination"]),
-                     Return([Nil()])]))
+                     Return([self._error_exit_code])]))
         body.append(If(*if_sections))
         self.exit_scope()
 
@@ -692,9 +939,15 @@ class CToPythonWrapper(Wrapper):
             The function which can be called from Python.
         """
         original_func = getattr(expr, 'original_function', expr)
-        func_name = self.scope.get_new_name(original_func.name+'_wrapper')
+        func_name = self.scope.get_new_name(expr.name+'_wrapper')
         func_scope = self.scope.new_child_scope(func_name)
         self.scope = func_scope
+
+        possible_class_base = expr.get_user_nodes((ClassDef,))
+        if possible_class_base:
+            class_dtype = possible_class_base[0].class_type
+        else:
+            class_dtype = None
 
         is_bind_c_function_def = isinstance(expr, BindCFunctionDef)
 
@@ -735,7 +988,7 @@ class CToPythonWrapper(Wrapper):
             func_args = [FunctionDefArgument(a) for a in self._get_python_argument_variables(python_args)]
             body = []
         else:
-            func_args, body = self._unpack_python_args(python_args)
+            func_args, body = self._unpack_python_args(python_args, class_dtype)
             func_args = [FunctionDefArgument(a) for a in func_args]
 
         # Get the results of the PyFunctionDef
@@ -812,7 +1065,8 @@ class CToPythonWrapper(Wrapper):
 
         self.exit_scope()
         for a in python_args:
-            self._python_object_map.pop(a)
+            if not a.bound_argument:
+                self._python_object_map.pop(a)
         for r in python_results:
             self._python_object_map.pop(r)
 
@@ -855,6 +1109,7 @@ class CToPythonWrapper(Wrapper):
         in_interface = len(expr.get_user_nodes(Interface)) > 0
 
         orig_var = getattr(expr, 'original_function_argument_variable', expr.var)
+        bound_argument = getattr(expr, 'wrapping_bound_argument', expr.bound_argument)
 
         if orig_var.is_ndarray:
             arg_var = orig_var.clone(self.scope.get_expected_name(orig_var.name), is_argument = False, memory_handling='alias')
@@ -886,14 +1141,19 @@ class CToPythonWrapper(Wrapper):
             python_cls_base = self.scope.find(dtype.name, 'classes', raise_if_missing = True)
             scope = python_cls_base.scope
             attribute = scope.find('instance', 'variables', raise_if_missing = True)
-            cast_type = Variable(dtype=self._python_object_map[dtype],
-                                name=self.scope.get_new_name(collect_arg.name),
-                                memory_handling='alias',
-                                cls_base = self.scope.find(dtype.name, 'classes', raise_if_missing = True))
-            self.scope.insert_variable(cast_type)
+            if expr.bound_argument:
+                cast_type = collect_arg
+                cast = []
+            else:
+                cast_type = Variable(dtype=self._python_object_map[dtype],
+                                    name=self.scope.get_new_name(collect_arg.name),
+                                    memory_handling='alias',
+                                    cls_base = self.scope.find(dtype.name, 'classes', raise_if_missing = True))
+                self.scope.insert_variable(cast_type)
+                cast = [AliasAssign(cast_type, PointerCast(collect_arg, cast_type))]
             c_res = attribute.clone(attribute.name, new_class = DottedVariable, lhs = cast_type)
-            cast = [AliasAssign(cast_type, PointerCast(collect_arg, cast_type)),
-                    AliasAssign(arg_var, PointerCast(c_res, arg_var))]
+            cast_c_res = PointerCast(c_res, orig_var)
+            cast.append(AliasAssign(arg_var, cast_c_res))
         elif arg_var.rank == 0:
             prec  = get_final_precision(orig_var)
             try :
@@ -916,12 +1176,12 @@ class CToPythonWrapper(Wrapper):
         if expr.has_default:
             check_func, err = self._get_check_function(collect_arg, orig_var, False)
             body.append(If( IfSection(check_func, cast),
-                        IfSection(PyccelIsNot(collect_arg, Py_None), [*err, Return([Nil()])])
+                        IfSection(PyccelIsNot(collect_arg, Py_None), [*err, Return([self._error_exit_code])])
                         ))
-        elif not in_interface:
+        elif not (in_interface or bound_argument):
             check_func, err = self._get_check_function(collect_arg, orig_var, True)
             body.append(If( IfSection(check_func, cast),
-                        IfSection(LiteralTrue(), [*err, Return([Nil()])])
+                        IfSection(LiteralTrue(), [*err, Return([self._error_exit_code])])
                         ))
         else:
             body.extend(cast)
@@ -1219,11 +1479,36 @@ class CToPythonWrapper(Wrapper):
 
         type_name = self.scope.get_new_name(f'Py{name}Type')
         docstring = expr.docstring
-        wrapped_class = PyClassDef(expr, struct_name, type_name, Scope(), docstring = docstring, class_type = dtype)
+        wrapped_class = PyClassDef(expr, struct_name, type_name, self.scope.new_child_scope(expr.name),
+                                   docstring = docstring, class_type = dtype)
+
+        orig_cls_dtype = expr.scope.parent_scope.cls_constructs[name]
 
         self._python_object_map[expr] = wrapped_class
-        self._python_object_map[expr.scope.parent_scope.cls_constructs[name]] = dtype
+        self._python_object_map[orig_cls_dtype] = dtype
 
         self.scope.insert_class(wrapped_class, name)
+        orig_scope = expr.scope
+
+        for f in expr.methods:
+            orig_f = getattr(f, 'original_function', f)
+            name = orig_f.name
+            python_name = orig_scope.get_python_name(name)
+            if python_name == '__del__':
+                wrapped_class.add_new_method(self._get_class_destructor(f, orig_cls_dtype, wrapped_class.scope))
+            elif python_name == '__init__':
+                wrapped_class.add_new_method(self._get_class_initialiser(f, orig_cls_dtype))
+            else:
+                wrapped_class.add_new_method(self._wrap(f))
+
+        for i in expr.interfaces:
+            for f in i.functions:
+                self._wrap(f)
+            wrapped_class.add_new_interface(self._wrap(i))
+
+        if isinstance(expr, BindCClassDef):
+            wrapped_class.add_alloc_method(self._get_class_allocator(orig_cls_dtype, expr.new_func))
+        else:
+            wrapped_class.add_alloc_method(self._get_class_allocator(orig_cls_dtype))
 
         return wrapped_class